+++ notifications_utils/field.py
@@ -3,13 +3,13 @@ import re
 from markupsafe import Markup
 from orderedset import OrderedSet
 
-from notifications_utils.columns import Columns
 from notifications_utils.formatters import (
     escape_html,
     strip_and_remove_obscure_whitespace,
     strip_html,
     unescaped_formatted_list,
 )
+from notifications_utils.insensitive_dict import Columns
 
 
 class Placeholder:
